Changed settings file handling to use LOVE's API
There was a problem after compiling where the program was not able to
create or access the settings file. Now it should not be a problem.
Might further change this to support love.filesystem.load(), where I can
use a valid lua file to contain the settings and load them as useable
variables natively instead of parsing them from strings.