Fixed crash if path leads to or past 0 in y/x
The limiter that prevents infinite loops will also catch if it goes past
the right/bottom bounds of the grid as well. Though, this will only
catch it after it has gone through mapHeight * mapWidth iterations
before failing. If this proves inefficient (it's instant as it is) or
unreliable, I'll just add checks for right/bottom grid bounds as well.