Fix for a few bugs regarding clicking out of grid
I had already put in the logic for determining if the grid is what was
clicked, I don't know why I didn't actually put it to use. The 1 pixel
margin between buttons could be clicked and it would set the start
position at 1, 1. The buttons could also be right clicked and they would
activate but not fire the click animation. They can now only be left
clicked. There wasn't technically anything buggy with right clicks on
the grid but i moved that within the logic for determining grid clicks
as well as it should have been there in the first place.