Fix for fresh install crash when opening settings
I had written the method for reading settings quite a while ago and I
hadn't ever come across this crash until building from source on a
different computer and opening settings immediately after running it for
the first time. When the settings file was being created, the read
method wasn't returning anything because it was running read(), which
was returning the settings to the parent read() but not to whatever
called that parent read(). Simple to overlook. Don't need to return
anything other than defaults when creating the settings file for the
first time, so that's how it's being handled now.