Added beginnings of toast class
Still not sure how I will handle the display of toasts. As it stands I
will have to add whichever class is calling a toast to
love.update/draw/etc if it's not there already to allow it to display
toasts, and I'll have to add callbacks during clicking for that calling
class to check if it had something clickable get clicked.