Change button instance init syntax
I had wanted a button that has its enabled state set on init to have a
syntax dictating explicitly that the enabled state of the button is
what was being set, but I eventually decided this was unnecessary.

Enabled state no longer needs to be passed as a table member and instead
to be passed as an optional fifth parameter of boolean value.