Refactor to massively cut back on linter warnings
- Main.lua functions have been moved up and declared local. These are
mostly related to the map building functionality that I intend to move
to a separate class so this won't likely matter soon

- Class loaders have been given underscores (largely because button
class loaders were being shadowed by love.mousepressed/mousereleased. It
wasn't a problem but the warnings were obnoxious and I didn't want to
make the change only for button imports).

- Local variables that overload broader scope vars of the same name have
been renamed. This cut back on a huge number of warnings.

As it stands right now, the only remaining warnings are for declared but
uninitialized class public fields, unused argument vars within love
methods, and empty if branches which I leave open in case I want to
stick logic in there.