Move map editing functionality to MapEditor class
I still need to clean up the mess I made in main.lua, which I intend to
work on tidying while at work. Out of time for now, but the MapEditor
class is fully operational and functions exactly as it did when the map
editing code was in the main program.