Rename some things, clean up code
Moving back to keeping love analogous functions named the same as their
love counterparts anywhere that I had strayed from this convention,
excluding with tooltips and toast because they are not drawn directly.
They are added to the view as invisible objects and set to display via a
separate method.