Make direction cell markers more tile-like
I feel it looks a lot better than simply coloring in the grid squares
and putting the marker on them. I also feel it makes it much easier to
follow the path on the editor with your eyes when designing it.

I may also add some sort of system that visibly highlights the valid
path so it can be seen if a lot of other invalid markers are placed
around it. Not really necessary when invalid markers aren't used
(because why would they unless you're just doodling) but it'd be a nice
indicator to go along with the export button becoming active that this
map is valid.